+++ src/components/Admin/Admin.tsx
@@ -1,13 +1,13 @@
 import {FC} from 'react'
 import {Admin as ReactAdmin, Resource} from 'react-admin'
 
-import {dataProvider} from '../../admin/dataProvider'
-import {authProvider} from '../../admin/tokenAuthProvider'
 import {CompetitionList} from './Competition/CompetitionList'
+import {dataProvider} from './dataProvider'
 import {PostCreate} from './Post/PostCreate'
 import {PostEdit} from './Post/PostEdit'
 import {PostList} from './Post/PostList'
 import {PostShow} from './Post/PostShow'
+import {authProvider} from './tokenAuthProvider'
 
 export const Admin: FC = () => (
   <ReactAdmin authProvider={authProvider} dataProvider={dataProvider}>
